refactor: begin matrix testing
