+++ docker/.env-8.8
@@ -1,0 +1,23 @@
+## Image versions ##
+# CAMUNDA_CONNECTORS_VERSION=0.23.2
+CAMUNDA_CONNECTORS_VERSION=8.6.0
+CAMUNDA_OPTIMIZE_VERSION=8.8.0-alpha1
+CAMUNDA_PLATFORM_VERSION=8.8.0-alpha1
+CAMUNDA_ZEEBE_VERSION=8.8.0-alpha1
+CAMUNDA_WEB_MODELER_VERSION=8.6.0
+ELASTIC_VERSION=8.9.0
+KEYCLOAK_SERVER_VERSION=22.0.3
+MAILPIT_VERSION=v1.7.0
+POSTGRES_VERSION=14.5-alpine
+HOST=localhost
+
+## Configuration ##
+# By default the zeebe api is public, when setting this to `identity` a valid zeebe client token is required
+ZEEBE_AUTHENTICATION_MODE=identity
+ZEEBE_CLIENT_ID=zeebe
+ZEEBE_CLIENT_SECRET=zecret
+
+# Set to 'true' to enable resource based authorizations for users and groups
+# This can be used to limit access for users or groups to view/update specific
+# processes and decisions in Operate and Tasklist
+RESOURCE_AUTHORIZATIONS_ENABLED=false
